Lazy caching for Sonar server version.
Lazying caching of SonarQube server version improves performance if the client code is interested in servers details except version. Example User wants to edit the servers list in Eclipse preference when SonarQube server is not reachable.
Version caching and it's parsing has been encapsulated to a new Java file 'Version'.