Project configuration setup improvement
SonaQube server association action and it's dialogue improvement when few of saved/persisted SonarQube servers are not reachable at given time. User can still be able to associated a workspace project to a SonarQube server if at-least one of SonarQube servers is reachable, with additional warning message shown on the association dialogue.