SLE-496 Unpack the plugins at installation time
The 'unpack' flag in the feature.xml file is not sufficient. See https://wiki.eclipse.org/Tycho/Packaging_Types#eclipse-feature. It says:

Limitation: The unpack attribute in the feature.xml is ignored. If you need a bundle to be installed in unpacked form, specify Eclipse-BundleShape: dir in the manifest of the bundle. If this is a third party bundle, it will mean repackaging it to include this Eclipse proprietary value.