Refresh a connection's label in the Bindings view after udpate
Prior to that the label was refreshed only when updating projects binding. So if a connection has no project bound, it could happen that the label was not refreshed (state was remaining 'Unknown'). It's not a new problem but one test started failing, maybe due to the engine startup/sync being slower