SLE-825: Share Connected Mode setup among contributors (#636)
* SLE-830: New EP for project hierarchy awareness

This will help as for example Maven projects with many modules are flat inside Eclipse.

* SLE-828: Users can share Connected Mode config

This allows the user to invoke it from the context menu on the project and on the SonarLint Bindings view.

* SLE-827: Handle the suggestConnection

TODO ...

* SLE-839: Handle files added / changed correctly

For SonarLint configuration files used for creating Connection Suggestions we have to work with Eclipse quirks on when a project is added and files are either changed or added.

* SLE-827: Handle the suggestConnection

More centralized logic and dialogs shown with more information.

* Feedback improvements

* SLE-841,SLE-45: Update LSP4J and use abstractions

Due to changes on SLCORE LSP4J must be updated from 0.21.1 to 0.22.0. These changes included the introduction of an abstraction layer of LSP4J types used and exposed to the clients.