SLE-861: Try to rely on Eclipse provided Java installation
Since 10.0 we bundle our on Java installation in order to keep SonarLint running smoothly on IDEs running on Java 16 or lower.
For power users we removed the friction already a bit by adding the ability to provide their own self-managed Java runtime.
To take this one step further we want to re-use the Java installation of Eclipse if possible, as power users providing their own self-managed Java runtime to SonarLint would do the same for Eclipse.