SLE-870: Rework TriggerType
Had to be adjusted to the analysis and tracking out of process. By doing so some dead code was removed as well.

Correctly disposing the AnalysisTracker in case of an exception in the analysis itself and before the actual IssuesMarkerUpdateJob.

Limited the usage of the TriggerType as only the information if markers are "on-the-fly" ones or not is required for updating the issue markers.