SLE-883: Stabilize unit tests
Both these tests would either both succeed or fail. This is now decoupled.

Additionally, the events can take longer, so we have to wait for them. Due to us not being able to use the Awaitility library here (due to constrains of building Maven against latest Eclipse release with Java 11) we have to mimic that behavior.