SLE-888: Split debug logs and IDE-specific traces (#687)
* SLE-888: Split debug logs and IDE-specific traces

Currently the debug logs will contain all the traces we use for Eclipse IDE specific debugging, like adaptations of different elements to SonarLint-specific classes or interaction with the extension points provided by SonarLint.

This is very useful information but does collide with the debug logs in may ways and is only necessary for Eclipse-specific debugging.

* SLE-888: PR feedback

Moved `trace(...)` to `traceIdeMessage(...)` in order to not confuse it with SLCORE tracing messages.

Updated the integration test to clarify that the IDE-specific logs work. Also, moved to Awaitility library.