SLE-893: Handle setting up Connected Mode automatically
When running Open in IDE (no matter of SonarQube or SonarCloud) and no connection / binding is established, we have to do it first. When the feature is then invoked afterwards, we have to await the connection to become ready (including analysis and branch matching) before we can proceed.
We want to provide the user with information and a more clear UX.