SLE-897: Consistency for SLCORE versions
In order to not ship two different versions, we have to have a consistency check between the versions we pass to the Maven "pom.xml" file and the Eclipse PDE target platform "commons.target".

Sadly it is not possible to only set the version once as the target platform cannot be extended with a property or reference another file or anything else.