SLE-920: Move EP to the UI plug-in
In order to not calculate SWT for all fragments (Sloop) of the CORE plug-in, we move it to the UI plug-in.

This will fix an error with the build as the SWT dependency couldn't be found for our "no-arch" fragment that is literally "every other OS/ARCH" combination except for the ones we have Sloop packages for.