SLE-900: EP for excluding unrelated files from indexing
An extension point is implemented that is used for narrowing down the focus of files that are actually part of a project and relevant to SonarLint.

This will speed up SonarLint and lower the memory footprint inside the IDE and for SLCORE.

To speed up operations on the IDE side, especially for importing projects and/or opening a workspace, caches were put in place for both the new extension point and files of a project as calculating both are very costly.

This also includes the exclusion of VCS files and narrowing the focus for Node.js related files as this wouldn't make sense to be put on sub-plugins.