SLE-948: Drop PyDev QOL integration (#747)
As the PyDev dependencies (i.e. their Eclipse Update Sites) are highly unreliable for local development and the build, this integration was removed.

It only featured some QOL additions like the syntax highlighting that is only available when working with PyDev. The logic to exclude Python virtual environment related paths from indexing is kept and moved to SonarLintUtils.