SLE-969: Create integration test with CDT
Using simple Makefile-based project for testing the C analysis and configuration in the sub-module.

Additionally, updated the CI components to provide all tools for using gcc/g++/make with Eclipse CDT.

The CI image now needs libwebkit2gtk-4.* as the underlying rendering component of the browser instances used in the SonarQube Rule Description view. Due to Eclipse 4.8 (the oldest target platform) requiring GTK3 - this is not available on the Linux machines on the CI - the RuleDescriptionViewTest was disabled on that axis!