SLE-1000: Don't run the analysis on files with no CDT information
When the analysis is started but CDT cannot provide information about (e.g. because it is not yet ready, ill configured project).

In this cases, firstly a NullPointerException was thrown, and secondly the analysis failed.

We mitigated that issue by "removing" (not excluding - to not confuse the terms) these files from the analysis and logging an error with enough information.

This way users can still analyze files but CFamily analysis will be skipped.

A unit test was fixed that was always succeeding but testing an incorrect behavior.