SLE-1098: Exclude buggy Eclipse PDE files
Eclipse PDE alongside Maven/Tycho has a bug when using Target Platform definitions and `product` definition files in a project, that a folder is created named `file:` that contains unpacked parts of plug-ins.
This triggers a lot of events every time something changes - it consumes a lot of bandwidth and should therefore be ignored.