SLE-1113: Rely on Eclipse Mylyn of released version (#837)
Don't rely on the latest Eclipse Mylyn version for the latest Java 17 based Eclipse IDE Target Platform.

Consequence is that the actual feature name of Mylyn is inconsistent and had to be updated in the `product` definition of the Connected Mode ITs against SonarQube Server (all running with Java 17).

The Standalone Mode ITs on the CI also don't run against Java 17 anymore as we are testing both the oldest and the newest supported version, this was a leftover of July last year when the Eclipse IDE moved from Java 17 to 21.