SLE-1133: Correctly exclude files from indexing
We currently exclude files from VCS, Node.js and Python virtual environments from indexing (and therefore the analysis) to optimize our plug-in.

We should do it correctly by checking for the actual files and folder and not whether an excluded identifier is part of the string of the absolute path.