SLE-1138: EP to load Connected Mode tokens
We currently rely only on Eclipse Equinox Security which is not available on some Eclipse-based IDEs. These users cannot work with Connected Mode then.

We added an extension point and moved our Equinox Security related logic to a default implementation so third-parties (or ourselves in the future) can rely on different options for storing the encrypted token.