SLE-1102: Implement monitoring via Sentry.io
On an early startup we initialize Sentry.io to catch as many exceptions as possible before the actual logging is initialized and available in the SonarQube Console.

The MonitoringService is also responsible for making sure that Sentry.io is enabled or disabled completely both for the IDE plug-in and SLOOP.

For uncaught exceptions in the IDE where we would normally see a "Problem Occurred" dialog we will also capture them and tag them specifically.