SLE-1157: Don't analyze when a quick fix was applied
This does not make sense due to the fact that when applying a quick fix the file is not saved and therefore the analysis runs on the pre-quick-fix-applied file.

This is a useless analysis. And in anticipation of the upcoming changes to the analysis triggering API of SonarLint Core this is removed to have less confusion.