SLE-1167: Remove analysis completion method and deprecate EP
Since the analysis is asynchronous nowadays, the "AnalyzeProjectJob#analysisCompleted" did not make sense anymore - as the analysis request was only handed over to SLCORE, but not finished.

The extension point method in the background did not make sense anymore and should be replaced on the implementing side (e.g. by a scheduled cleanup).
The related logic for that was therefore deprecated to be removed with the next major release.