Handle the failure to open the Welcome page more gracefully
When using a Eclipse-based application that does not provide the Welcome view at all or not the intro page we extend and that is opened by default by the Welcome page, we should not throw a unhandled runtime exception. In those and other cases, just not open the Welcome page.