Workaround for duplicate markers on the same resource
This seems to be a rare corner case for when switching a branch and the new branch has the same issue on the file, a duplicate marker is found.
In case there is a duplicate, we take into account the "new" one and proceed as in the following logic for all existing markers there is no new one created.